Refactor random pokemon to separate component
